+++ package.json
@@ -12,7 +12,8 @@
   "dependencies": {
     "@picocss/pico": "^2.0.6",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router": "^6.22.2"
   },
   "devDependencies": {
     "@types/react": "^18.2.56",
