+++ src/Components/TodoItemFormReducer.jsx
@@ -16,55 +16,8 @@ const progressStates = ["To Do", "In Progress", "Done"];
 
 const priorities = ["Low", "Medium", "High"];
 
-const TodoItemFormReducer = () => {
-	const addNewTodo = (state, action) => {
-		switch (action.type) {
-			case "form-submit": {
-				action.payload.preventDefault();
-				console.log("got a form submit", action.payload);
-				const formData = new FormData(action.payload.target.form);
-
-				const vals = {};
-				for (const key of formData.keys()) {
-					vals[key] = formData.get(key);
-				}
-				const formatDate = () => {
-					const dateNow = new Date();
-					return `${dateNow.getFullYear()}-${(dateNow.getMonth() + 1)
-						.toString()
-						.padStart(2, "0")}-${dateNow
-						.getDate()
-						.toString()
-						.padStart(2, "0")}`;
-				};
-				console.log(vals);
-				//setTodoItems([...todoItems, { ...state, ...vals }]);
-				return {
-					...state,
-					timestamp: new Date(),
-					createdDate: formatDate(),
-					...vals,
-				};
-			}
-			default: {
-				console.log("state", state);
-				console.log("action", action);
-			}
-		}
-	};
-
-	const [state, dispatch] = useReducer(addNewTodo, {
-		id: crypto.randomUUID(),
-		title: "",
-		description: "",
-		assignee: "",
-		status: "To Do",
-		priority: "Low",
-		dueDate: "",
-		createdDate: "",
-	});
-
-	console.log(reducer);
+const TodoItemFormReducer = ({ dispatch, state }) => {
+	//console.log(dispatch);
 	return (
 		<form style={formStyle}>
 			<label htmlFor="todo-item-title">Title</label>
@@ -73,8 +26,6 @@ const TodoItemFormReducer = () => {
 				name="title"
 				type="text"
 				placeholder="Task title..."
-				// value={state.title}
-				//onChange={(e) => setTitle(e.target.value)}
 			/>
 
 			<label htmlFor="todo-item-description">Description</label>
@@ -102,7 +53,7 @@ const TodoItemFormReducer = () => {
 				{priorities.map((element) => {
 					const nameToValue = element.toLowerCase().replaceAll(" ", "-");
 					return (
-						<option key={nameToValue} value={nameToValue}>
+						<option key={nameToValue} value={element}>
 							{element}
 						</option>
 					);
@@ -117,7 +68,7 @@ const TodoItemFormReducer = () => {
 				{teamMembers.map((element) => {
 					const nameToValue = element.toLowerCase().replaceAll(" ", "-");
 					return (
-						<option key={nameToValue} value={nameToValue}>
+						<option key={nameToValue} value={element}>
 							{element}
 						</option>
 					);
