+++ src/Pages/Dashboard.jsx
@@ -1,0 +1,4 @@
+const Dashboard = () => {
+	return <div>Dashboard</div>;
+};
+export default Dashboard;
